Stream results from JDBC connection
By default, JDBC mariaDB implementation is fetching all the rows on memory while obtaining a query ResultSet.

Adding a specific fetch size of 100 on the prepared statement we can stream large amounts of data directly from gitbase to a parquet file without memory problems.

Also found a problem related with rule registration on SparkSession using the spark shell. Adding the workaround on the readme example to avoid problems with people trying GSC on the shell directly.

Signed-off-by: Antonio Jesus Navarro Perez <antnavper@gmail.com>