fix quoting of identifiers in gitbase dialect for JDBC
Fixes #86

When using MariaDB JDBC dialect, the identifiers are quoted using double quotes, which
Gitbase interprets as a string, not a quoted identifier (as does MySQL).
This adds quoteIdentifier to the GitbaseDialect to quote using backticks and registers
the dialect globally so it's used instead of the default MariaDB one.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>